@@ -1,25 +1,13 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <Nullable>enable</Nullable>
-        <LangVersion>9</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="9.0.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.2" />
-        <PackageReference Include="Serilog.AspNetCore" Version="3.4.0" />
-        <PackageReference Include="Serilog.Exceptions" Version="6.0.0" />
-        <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="5.0.2" />
+      <ProjectReference Include="..\Navigator.Providers.Telegram\Navigator.Providers.Telegram.csproj" />
+      <ProjectReference Include="..\Navigator\Navigator.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-        <ProjectReference Include="..\Navigator.Abstractions.Extensions\Navigator.Abstractions.Extensions.csproj" />
-        <ProjectReference Include="..\Navigator\Navigator.csproj" />
-        <ProjectReference Include="..\Navigator.Extensions.Actions\Navigator.Extensions.Actions.csproj" />
-        <ProjectReference Include="..\Navigator.Extensions.Shipyard\Navigator.Extensions.Shipyard.csproj" />
-    </ItemGroup>
-    
 </Project>
