@@ -1,27 +1,23 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using MediatR;
-using Navigator.Abstractions;
-using Navigator.Abstractions.Extensions;
-using Navigator.Extensions.Actions;
-using Navigator.Extensions.Store.Abstractions.Extensions;
-using Navigator.Samples.Echo.Entity;
-using Newtonsoft.Json;
+using Navigator.Actions;
+using Navigator.Context;
+using Navigator.Providers.Telegram;
+using Telegram.Bot;
 
-namespace Navigator.Samples.Echo.Actions
+namespace Navigator.Samples.Echo.Actions;
+
+public class EchoActionHandler : ActionHandler<EchoAction>
 {
-    public class EchoActionHandler : ActionHandler<EchoAction>
+    public EchoActionHandler(INavigatorContextAccessor navigatorContextAccessor) : base(navigatorContextAccessor)
     {
-        public EchoActionHandler(INavigatorContext ctx) : base(ctx)
-        {
-        }
+    }
 
-        public override async Task<Unit> Handle(EchoAction request, CancellationToken cancellationToken)
-        {
-            await Ctx.Client.SendTextMessageAsync(Ctx.GetTelegramChat(), request.MessageToEcho,
-                cancellationToken: cancellationToken);
+    public override async Task<Status> Handle(EchoAction action, CancellationToken cancellationToken)
+    {
+        await this.GetTelegramClient().SendTextMessageAsync(this.GetTelegramChat().Id, 
+            action.MessageToEcho, cancellationToken: cancellationToken);
 
-            return Unit.Value;
-        }
+        return Success();
     }
 }
